defenses.py
```python
from math_formulas import DEFENSE_FORMULAS
from utils import calculateDistance, convertFormulaToPath
from constants import BASE_HEALTH, UPGRADE_COSTS

class Defense:
    def __init__(self, formula, position, health):
        self.formula = formula
        self.position = position
        self.health = health
        self.path = convertFormulaToPath(formula)

    def hit(self, damage):
        self.health -= damage
        if self.health <= 0:
            self.destroy()

    def destroy(self):
        # Placeholder for destruction logic
        pass

    def draw(self, visualization):
        visualization.draw_path(self.path)

class DefenseManager:
    def __init__(self):
        self.defenses = []

    def create_defense(self, formula_type, position):
        if formula_type in DEFENSE_FORMULAS:
            formula = DEFENSE_FORMULAS[formula_type]
            health = BASE_HEALTH  # Initial health can be a function of the formula complexity
            defense = Defense(formula, position, health)
            self.defenses.append(defense)
            return defense
        else:
            raise ValueError("Unknown defense formula type")

    def upgrade_defense(self, defense, upgrade_type):
        if upgrade_type in UPGRADE_COSTS:
            cost = UPGRADE_COSTS[upgrade_type]
            # Placeholder for upgrade logic, e.g., increase health, change formula, etc.
            pass
        else:
            raise ValueError("Unknown upgrade type")

    def handle_projectile_impact(self, projectile):
        for defense in self.defenses:
            if calculateDistance(defense.position, projectile.position) <= projectile.radius:
                defense.hit(projectile.damage)
                if defense.health <= 0:
                    self.defenses.remove(defense)

    def draw_defenses(self, visualization):
        for defense in self.defenses:
            defense.draw(visualization)
```