tests/test_defenses.py
```python
import unittest
from defenses import Defense
from math_formulas import circle_formula, line_formula
from game_logic import GameLogic

class TestDefenses(unittest.TestCase):
    def setUp(self):
        self.game_logic = GameLogic()
        self.base_health = self.game_logic.base_health

    def test_create_circle_defense(self):
        radius = 5
        defense = Defense(circle_formula(radius))
        self.assertIsNotNone(defense)
        self.assertEqual(defense.formula, circle_formula(radius))

    def test_create_line_defense(self):
        slope = 1
        y_intercept = 0
        defense = Defense(line_formula(slope, y_intercept))
        self.assertIsNotNone(defense)
        self.assertEqual(defense.formula, line_formula(slope, y_intercept))

    def test_defense_health_initialization(self):
        defense = Defense(circle_formula(5))
        self.assertEqual(defense.health, self.game_logic.initial_defense_health)

    def test_defense_damage(self):
        defense = Defense(circle_formula(5))
        initial_health = defense.health
        defense.take_damage(PROJECTILE_DAMAGE)
        self.assertEqual(defense.health, initial_health - PROJECTILE_DAMAGE)

    def test_defense_destroyed(self):
        defense = Defense(circle_formula(5))
        defense.take_damage(defense.health)  # Apply damage equal to its health
        self.assertTrue(defense.is_destroyed())

    def test_upgrade_defense(self):
        defense = Defense(circle_formula(5))
        initial_health = defense.health
        self.game_logic.apply_upgrade(defense, 'health')
        self.assertGreater(defense.health, initial_health)

    def test_defense_intersects_projectile(self):
        defense = Defense(line_formula(1, 0))
        projectile_position = (3, 3)  # Should be on the line y=x
        self.assertTrue(self.game_logic.is_collision(defense, projectile_position))

    def test_defense_does_not_intersect_projectile(self):
        defense = Defense(line_formula(1, 0))
        projectile_position = (3, 4)  # Should not be on the line y=x
        self.assertFalse(self.game_logic.is_collision(defense, projectile_position))

if __name__ == '__main__':
    unittest.main()
```